src:ticket:932

+++ ROMS/Bin/cbuild_roms.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 #
 # git $Id$
-# svn $Id: cbuild_roms.sh 1151 2023-02-09 03:08:53Z arango $
+# svn $Id: cbuild_roms.sh 1160 2023-03-25 21:13:13Z arango $
 #::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
 # Copyright (c) 2002-2023 The ROMS/TOMS Group                           :::
 #   Licensed under a MIT/X style license                                :::
@@ -40,6 +40,8 @@
 #                                                                       :::
 #    -noclean    Do not clean already compiled objects                  :::
 #                                                                       :::
+#    -v          Compile in verbose mode (VERBOSE=1)                    :::
+#                                                                       :::
 #::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
 
 export which_MPI=openmpi                       # default, overwritten below
@@ -47,6 +49,7 @@ export which_MPI=openmpi                       # default, overwritten below
 parallel=0
 clean=1
 dprint=0
+verbose=0
 
 export MY_CPP_FLAGS=
 
@@ -73,6 +76,11 @@ do
       shift
       ;;
 
+    -v )
+      shift
+      verbose=1
+      ;;
+
     -noclean )
       shift
       clean=0
@@ -91,6 +99,8 @@ do
       echo "              For example:  cbuild_roms.sh -p FFLAGS"
       echo ""
       echo "-noclean    Do not clean already compiled objects"
+      echo ""
+      echo "-v          Compile in verbose mode"
       echo ""
       exit 1
       ;;
@@ -222,24 +232,26 @@ fi
 
 # Create the build directory specified above and change into it.
 
-if [ -d ${SCRATCH_DIR} ]; then
-  if [ $clean -eq 1 ]; then
-    rm -rf ${SCRATCH_DIR}
-    mkdir ${SCRATCH_DIR}
-    cd ${SCRATCH_DIR}
+if [ $dprint -eq 0 ]; then
+  if [ -d ${SCRATCH_DIR} ]; then
+    if [ $clean -eq 1 ]; then
+      rm -rf ${SCRATCH_DIR}
+      mkdir ${SCRATCH_DIR}
+      cd ${SCRATCH_DIR}
+    else
+      cd ${SCRATCH_DIR}
+    fi
   else
-    cd ${SCRATCH_DIR}
-  fi
-else
-  if [ $clean -eq 1 ]; then
-    mkdir ${SCRATCH_DIR}
-    cd ${SCRATCH_DIR}
-  else
-    echo "-noclean option activated when the build directory didn't exist"
-    echo "creating the directory and disabling -noclean"
-    clean=1
-    mkdir ${SCRATCH_DIR}
-    cd ${SCRATCH_DIR}
+    if [ $clean -eq 1 ]; then
+      mkdir ${SCRATCH_DIR}
+      cd ${SCRATCH_DIR}
+    else
+      echo "-noclean option activated when the build directory didn't exist"
+      echo "creating the directory and disabling -noclean"
+      clean=1
+      mkdir ${SCRATCH_DIR}
+      cd ${SCRATCH_DIR}
+    fi
   fi
 fi
 
@@ -419,9 +431,17 @@ if [ $dprint -eq 1 ]; then
   echo $debug:"${!debug}"
 else
   if [ $parallel -eq 1 ]; then
-    make $NCPUS
+    if [ $verbose -eq 1 ]; then
+      make VERBOSE=1 $NCPUS
+    else
+      make $NCPUS
+    fi
   else
-    make
+    if [ $verbose -eq 1 ]; then
+      make VERBOSE=1
+    else
+      make
+    fi
   fi
   make install
 fi
